requirements.txt
Django==3.2.6
djangorestframework==3.14.0
Pillow==10.0.0
django-cors-headers==4.3.0

